+++ demo/src/constants.js
@@ -1,0 +1,29 @@
+export const MAX_CHANNELS = 6;
+export const DEFAULT_VIEW_STATE = { zoom: -5.5, target: [30000, 10000, 0] };
+export const DEFAULT_OVERVIEW = {
+  margin: 25,
+  scale: 0.15,
+  position: 'bottom-left'
+};
+export const COLORMAP_OPTIONS = [
+  'viridis',
+  'greys',
+  'magma',
+  'jet',
+  'hot',
+  'bone',
+  'copper',
+  'summer',
+  'density',
+  'inferno'
+];
+export const COLOR_PALLETE = [
+  [0, 0, 255],
+  [0, 255, 0],
+  [255, 0, 0],
+  [255, 255, 0],
+  [255, 128, 0],
+  [255, 0, 255],
+  [0, 255, 255],
+  [255, 255, 255]
+];