+++ CHANGELOG.md
@@ -4,11 +4,21 @@
 
 ### Added
 
+### Changed
+
+## 0.1.5
+
+### Added
+
+- Add `loaderSelection` as prop for `VivViewerLayer`. This parameter is passed to `loader.getTile` and `loader.getRaster` in addtion to `x`, `y` and `z`.
 - Display warning for Safari users
 
 ### Changed
 
 - Refactored VivViewer to take in arbitrary views/layers
+- Refactored demo to allow various channels and change selection
+- Use reducer for handling channel state in Demo
+- Use Material UI for demo components, and showcase all viv features
 - Update the IMAGE_RENDERING doc
 
 ## 0.1.4